Remove now-unused fixtures and move functional-test-only fixtures
This tidies up the main conftest file a lot.

Some "page" fixtures are now only used by functional tests so they've
been moved to functional/conftest. The next step is to remove these
altogether nd get the functional tests using FactoryBoy too.
